[EXOP-10] populate drop downs on app startup and call separately
